+++ CHANGELOG.md
@@ -6,7 +6,8 @@ Changelog](https://keepachangelog.com/en/1.0.0/), and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Features
+### Changed
+- Clean request parameters serialization.
 - Return authorization and discovery URIs as URI maps instead of URI strings.
 ### Fixed
 - Typo in the client object (disovery instead of discovery).